(animedrv) add anime.h
